Fix importing packages in Lib/site-packages
Suppose you have the following package (or folder) inside
Lib/site-packages/:

mypackage/
    __init__.py   -> with only a print(“On init”)
    hello.py        -> with only a print(“On hello”)

The most common way to import my package.hello is:

“from mypackage import hello”.

However, this imports correctly mypackage/__init__.py (printing the “On
init”), but then it is not able to import hello.py. This happens
because a flag that says that “mypackage” is a real package is not set
to true. Then, the importing code does not continue to the importing of
“hello.py”.

This patch sets the $package flag of a package in site-packages to
‘true’ when Brython was successfully able to import a
packagename/__init__.py file.

It fixes the problem and is working fine on Firefox 32, Chrome 38 and
Safari 7.1.