Fixes issue #56 (number of replacements in string)
Brython is not respecting the ‘count’ argument to replace only a few
occurrences on a string. This patch fixes that by implementing a loop
instead of using Regular Expressions.

Regular expressions are slow in current browser implementations, and
this loop is at least 400% faster than using Regex in pure javascript
as can be seen in this jsperf:
http://jsperf.com/string-replace-comparison.