Feature: Start an implementation of continuous performance benchmarking.
The benchmarking is run using the python script www/speed/perf.py.
This script runs all (most) files in the benchmark subdirectory
in a browser, in regular cpython and additionally runs the octane
js benchmark in the browser. Each benchmark is run a configurable
number of times and the average, min and max runtimes are printed
as well as standard deviation & unbiased standard deviation.
These results are then printed and, optionally, saved to a sqlite3
database (not yet implemented), file, or pushed to plot.ly for
plotting (untested).

The script has some options (run it with --help to see them)
allowing one to skip some benchmarks, etc.

It is intended, that this script will also be run in the travis-ci,
so that each commit will be benchmarked and the results sent
to plot.ly (or some other, possibly self-hosted, service...)