Fix equality comparison for types (Fixes #749)
The `__eq__` method of a class should not be used when comparing two
classes (it should only be used when comparing instances of the class).
This patch replaces instances of

```javascript
_b_.getattr(objA, '__eq__')(objB)
```

which were used to test equality of `objA`, `objB` with

```javascript
$B.rich_comp("__eq__", objA, objB)
```

which is modified to properly handle the case when `objA` and `objB` are
types.