clean: explicitly name the (optional) third argument of $transition
The function $transition in py2js.js can optionally receive a third
argument (e.g. if the token is an operator, the third argument specifies
what operator it is). This commit changes the function signature to
explicitly show this third parameter and also replaces the use of the
"arguments" array inside the function.